requirements.txt
discord_py_interactions==4.3.1
python-dotenv==0.21.0
